fix(fused_rgdr_packed_decode): add explicit uint8_t cast for Mul repeat stride params
-Wc++11-narrowing in CANN 9.0.0 treats narrowing from uint32_t to uint8_t
in initializer list as error. rs (dkA / DIMS_PER_BLOCK) is uint32_t,
while Mul expects uint8_t repeat strides. Add static_cast<uint8_t> to
resolve the narrowing error for both S@K and S@Q MatVecMul calls.

+++ csrc/attention/fused_rgdr_packed_decode/op_kernel/fused_rgdr_packed_decode.h
@@ -321,7 +321,7 @@ __aicore__ inline void KernelFusedRgd<S>::Process() {
                     for (uint32_t k = 0; k < DK; k += FOLD_HALF) {
                         uint64_t mask = (k + FOLD_HALF <= DK) ? FOLD_HALF : (DK - k);
                         Mul(stP[k], stB[k], kB[hki * dkA + k], mask, (uint64_t)curV,
-                            {1, 1, 1, rs, rs, 0});
+                            {1, 1, 1, static_cast<uint8_t>(rs), static_cast<uint8_t>(rs), 0});
                     }
                 }
                 PipeBarrier<PIPE_V>();
@@ -366,7 +366,7 @@ __aicore__ inline void KernelFusedRgd<S>::Process() {
                     for (uint32_t k = 0; k < DK; k += FOLD_HALF) {
                         uint64_t mask = (k + FOLD_HALF <= DK) ? FOLD_HALF : (DK - k);
                         Mul(stP[k], stB[k], qtB[hki * dkA + k], mask, (uint64_t)curV,
-                            {1, 1, 1, rs, rs, 0});
+                            {1, 1, 1, static_cast<uint8_t>(rs), static_cast<uint8_t>(rs), 0});
                     }
                 }
                 PipeBarrier<PIPE_V>();
